+++ code/cell.compile.web/src/node/bundle/index.ts
@@ -1,0 +1,1 @@
+export * from './bundle';
